+++ packages/taro-cli/src/h5/index.ts
@@ -23,16 +23,19 @@ import {
   recursiveMerge,
   replaceAliasPath,
   resolveScriptPath,
-  checkCliAndFrameworkVersion
-} from '../util'
+  npm as npmProcess,
+  processTypeEnum,
+  PROJECT_CONFIG,
+  REG_SCRIPTS,
+  REG_TYPESCRIPT
+} from '@tarojs/helper'
 import {
   convertAstExpressionToVariable as toVar,
   convertObjectToAstExpression as objToAst,
   convertSourceStringToAstExpression as toAst
 } from '../util/astConvert'
-import { BUILD_TYPES, processTypeEnum, PROJECT_CONFIG, REG_SCRIPTS, REG_TYPESCRIPT } from '../util/constants'
-import * as npmProcess from '../util/npm'
-import { IBuildOptions, IOption } from '../util/types'
+import { checkCliAndFrameworkVersion } from '../util'
+import { IBuildOptions, IOption, IBuildHooks } from '../util/types'
 import {
   APIS_NEED_TO_APPEND_THIS,
   deviceRatioConfigName,
@@ -92,7 +95,7 @@ class Compiler {
   constructor (public appPath: string, entryFile?: string, isUi?: boolean) {
     const projectConfig = recursiveMerge({
       h5: defaultH5Config
-    }, require(path.join(appPath, PROJECT_CONFIG))(merge))
+    }, require(resolveScriptPath(path.join(appPath, PROJECT_CONFIG)))(merge))
     this.projectConfig = projectConfig
     const sourceDir = projectConfig.sourceRoot || CONFIG.SOURCE_DIR
     this.sourceRoot = sourceDir
@@ -191,7 +194,7 @@ class Compiler {
     return Promise.all(readPromises)
   }
 
-  async buildDist ({ watch, port }: IBuildOptions) {
+  async buildDist ({ watch, port }: IBuildOptions, { modifyWebpackChain, modifyBuildAssets, onBuildFinish }: IBuildHooks) {
     const isMultiRouterMode = get(this.h5Config, 'router.mode') === 'multi'
     const entryFileName = this.entryFileName
     const projectConfig = this.projectConfig
@@ -231,7 +234,7 @@ class Compiler {
       designWidth: projectConfig.designWidth,
       entry: merge(defaultEntry, h5Config.entry),
       env: {
-        TARO_ENV: JSON.stringify(BUILD_TYPES.H5)
+        TARO_ENV: JSON.stringify('h5')
       },
       isWatch: !!watch,
       outputRoot: outputDir,
@@ -241,7 +244,10 @@ class Compiler {
       sass: projectConfig.sass,
       plugins: projectConfig.plugins,
       port,
-      sourceRoot
+      sourceRoot,
+      modifyWebpackChain,
+      modifyBuildAssets,
+      onBuildFinish
     })
 
     const webpackRunner = await npmProcess.getNpmPkg('@tarojs/webpack-runner', this.appPath)
@@ -1459,14 +1465,14 @@ class Compiler {
 
 export { Compiler }
 
-export async function build (appPath: string, buildConfig: IBuildOptions) {
-  process.env.TARO_ENV = BUILD_TYPES.H5
-  await checkCliAndFrameworkVersion(appPath, BUILD_TYPES.H5)
+export async function build (appPath: string, buildConfig: IBuildOptions, buildHooks: IBuildHooks) {
+  process.env.TARO_ENV = 'h5'
+  await checkCliAndFrameworkVersion(appPath, 'h5')
   const compiler = new Compiler(appPath)
   await compiler.clean()
   await compiler.buildTemp()
   if (compiler.h5Config.transformOnly !== true) {
-    await compiler.buildDist(buildConfig)
+    await compiler.buildDist(buildConfig, buildHooks)
   }
   if (buildConfig.watch) {
     compiler.watchFiles()