+++ packages/taro-mini-runner/src/webpack/chain.ts
@@ -10,13 +10,6 @@ import { mapKeys, pipe } from 'lodash/fp'
 import * as UglifyJsPlugin from 'uglifyjs-webpack-plugin'
 import * as webpack from 'webpack'
 import { PostcssOption, ICopyOptions, IPostcssOption } from '@tarojs/taro/types/compile'
-import chalk from 'chalk'
-
-import { getPostcssPlugins } from './postcss.conf'
-
-import MiniPlugin from '../plugins/MiniPlugin'
-import { IOption } from '../utils/types'
-import { recursiveMerge, isNodeModule, resolveScriptPath } from '../utils'
 import {
   REG_SASS,
   REG_LESS,
@@ -25,21 +18,18 @@ import {
   REG_MEDIA,
   REG_FONT,
   REG_IMAGE,
-  BUILD_TYPES,
   REG_SCRIPTS,
   REG_TEMPLATE,
-  MINI_APP_FILES
-} from '../utils/constants'
+  recursiveMerge,
+  isNodeModule,
+  resolveScriptPath,
+  chalk
+} from '@tarojs/helper'
 
-const globalObjectMap = {
-  [BUILD_TYPES.WEAPP]: 'wx',
-  [BUILD_TYPES.ALIPAY]: 'my',
-  [BUILD_TYPES.SWAN]: 'swan',
-  [BUILD_TYPES.QQ]: 'qq',
-  [BUILD_TYPES.TT]: 'tt',
-  [BUILD_TYPES.JD]: 'jd',
-  [BUILD_TYPES.QUICKAPP]: 'global'
-}
+import { getPostcssPlugins } from './postcss.conf'
+
+import MiniPlugin from '../plugins/MiniPlugin'
+import { IOption } from '../utils/types'
 
 const defaultUglifyJsOption = {
   keep_fnames: true,
@@ -163,6 +153,9 @@ export const getMiniPlugin = args => {
 
 export const getModule = (appPath: string, {
   sourceDir,
+  fileType,
+  isBuildQuickapp,
+  isUseComponentBuildPage,
 
   designWidth,
   deviceRatio,
@@ -183,7 +176,6 @@ export const getModule = (appPath: string, {
   alias,
   nodeModulesPath
 }) => {
-  const isQuickapp = buildAdapter === BUILD_TYPES.QUICKAPP
   const postcssOption: IPostcssOption = postcss || {}
 
   const cssModuleOptions: PostcssOption.cssModules = recursiveMerge({}, defaultCssModuleOption, postcssOption.cssModules)
@@ -224,7 +216,7 @@ export const getModule = (appPath: string, {
     include?;
     use;
   }[] = [{
-    use: isQuickapp ? [cssLoader, quickappStyleLoader] : [cssLoader]
+    use: isBuildQuickapp ? [cssLoader, quickappStyleLoader] : [cssLoader]
   }]
 
   const compileExclude = compile.exclude || []
@@ -256,6 +248,7 @@ export const getModule = (appPath: string, {
     {
       ident: 'postcss',
       plugins: getPostcssPlugins(appPath, {
+        isBuildQuickapp,
         designWidth,
         deviceRatio,
         postcssOption
@@ -264,7 +257,8 @@ export const getModule = (appPath: string, {
   ])
   const sassLoader = getSassLoader([{
     sourceMap: true,
-    implementation: sass
+    implementation: sass,
+    outputStyle: 'expanded'
   }, sassLoaderOption])
   const lessLoader = getLessLoader([{ sourceMap: enableSourceMap }, lessLoaderOption])
 
@@ -286,12 +280,12 @@ export const getModule = (appPath: string, {
     buildAdapter,
     constantsReplaceList: parsedConstantsReplaceList,
     sourceDir,
-    nodeModulesPath
+    isBuildQuickapp,
+    nodeModulesPath,
+    isUseComponentBuildPage
   }])
 
-  const miniTemplateLoader = getMiniTemplateLoader([{
-    buildAdapter
-  }])
+  const miniTemplateLoader = getMiniTemplateLoader([])
 
   let scriptsLoaderConf = {
     test: REG_SCRIPTS,
@@ -348,7 +342,7 @@ export const getModule = (appPath: string, {
       test: REG_TEMPLATE,
       use: [getFileLoader([{
         useRelativePath: true,
-        name: `[path][name]${MINI_APP_FILES[buildAdapter].TEMPL}`,
+        name: `[path][name]${fileType.templ}`,
         context: sourceDir
       }]), miniTemplateLoader]
     },
@@ -360,7 +354,7 @@ export const getModule = (appPath: string, {
           useRelativePath: true,
           context: sourceDir,
           ...mediaUrlLoaderOption,
-          limit: isQuickapp ? false : mediaUrlLoaderOption.limit
+          limit: isBuildQuickapp ? false : mediaUrlLoaderOption.limit
         }])
       }
     },
@@ -372,7 +366,7 @@ export const getModule = (appPath: string, {
           useRelativePath: true,
           context: sourceDir,
           ...fontUrlLoaderOption,
-          limit: isQuickapp ? false : fontUrlLoaderOption.limit
+          limit: isBuildQuickapp ? false : fontUrlLoaderOption.limit
         }])
       }
     },
@@ -384,7 +378,7 @@ export const getModule = (appPath: string, {
           useRelativePath: true,
           context: sourceDir,
           ...imageUrlLoaderOption,
-          limit: isQuickapp ? false : imageUrlLoaderOption.limit
+          limit: isBuildQuickapp ? false : imageUrlLoaderOption.limit
         }])
       }
     }
@@ -436,13 +430,13 @@ export const getEntry = ({
   }
 }
 
-export function getOutput (appPath: string, [{ outputRoot, publicPath, buildAdapter, isBuildPlugin }, customOutput]) {
+export function getOutput (appPath: string, [{ outputRoot, publicPath, globalObject }, customOutput]) {
   return {
     path: path.join(appPath, outputRoot),
     publicPath,
     filename: '[name].js',
     chunkFilename: '[name].js',
-    globalObject: globalObjectMap[buildAdapter],
+    globalObject,
     ...customOutput
   }
 }