+++ packages/taro-mobx-common/src/store.js
@@ -1,8 +1,12 @@
 let store = {}
+const appGlobal = global || {}
+const globalRef = Object.getPrototypeOf(appGlobal) || appGlobal
 
 export function getStore () {
   if (process.env.TARO_ENV === 'alipay') {
     return my.taroMobxStore || {}
+  } else if (process.env.TARO_ENV === 'quickapp') {
+    return globalRef.taroMobxStore || {}
   }
   return store
 }
@@ -10,6 +14,8 @@ export function getStore () {
 export function setStore (arg) {
   if (process.env.TARO_ENV === 'alipay') {
     my.taroMobxStore = arg
+  } else if (process.env.TARO_ENV === 'quickapp') {
+    globalRef.taroMobxStore = arg
   } else {
     store = arg
   }