+++ packages/taro-rn-runner/src/loaders/utils/processAst.ts
@@ -16,7 +16,7 @@ let componentClassName // get app.js class name
 const providerComponentName = 'Provider'
 const taroComponentsRNProviderName = 'TCRNProvider'
 const setStoreFuncName = 'setStore'
-const routerImportDefaultName = 'TaroRouter'
+// const routerImportDefaultName = 'TaroRouter'
 const DEVICE_RATIO = 'deviceRatio'
 
 const taroApis = [
@@ -576,7 +576,9 @@ export function processAst ({
             const parentPath = astPath.parentPath
 
             if (t.isMemberExpression(callee)) {
+              // @ts-ignore
               const object = callee.object as t.Identifier
+              // @ts-ignore
               const property = callee.property as t.Identifier
               if (object.name === taroImportDefaultName && property.name === 'render') {
                 astPath.remove()