+++ packages/taro-weapp/src/lifecycle.js
@@ -189,8 +189,8 @@ function doUpdate (component, prevProps, prevState) {
   }
 
   const cb = function () {
+    invokeEffects(component)
     if (__mounted) {
-      invokeEffects(component)
       if (component['$$refs'] && component['$$refs'].length > 0) {
         component['$$refs'].forEach(ref => {
           // 只有 component 类型能做判断。因为 querySelector 每次调用都一定返回 nodeRefs，无法得知 dom 类型的挂载状态。
@@ -229,6 +229,7 @@ function doUpdate (component, prevProps, prevState) {
   }
   if (Object.keys(dataDiff).length === 0) {
     cb()
+    invokeEffects(component)
   } else {
     component.$scope.setData(dataDiff, cb)
   }