+++ packages/taro/src/native-apis.js
@@ -50,7 +50,16 @@ const onAndSyncApis = {
   onAudioInterruptionEnd: true,
   onAudioInterruptionBegin: true,
   onLocationChange: true,
-  offLocationChange: true
+  offLocationChange: true,
+
+  onLocalServiceResolveFail: true,
+  onLocalServiceLost: true,
+  onLocalServiceFound: true,
+  onLocalServiceDiscoveryStop: true,
+  offLocalServiceResolveFail: true,
+  offLocalServiceLost: true,
+  offLocalServiceFound: true,
+  offLocalServiceDiscoveryStop: true
 }
 const noPromiseApis = {
   // 媒体
@@ -241,6 +250,9 @@ const otherApis = {
   setBackgroundTextStyle: true,
   getSelectedTextRange: true,
   hideHomeButton: true,
+
+  stopLocalServiceDiscovery: true,
+  startLocalServiceDiscovery: true,
 
   // 第三方平台
   getExtConfig: true,